Another Meh Update
I've changed just about every file in this update but it was just an adjustment to my PDO call. The problem with the parsing has been fixed by wrapping my fetch operations in an ajax call (I forgot) and now we're sort of good again. None of the php files appear to be functioning anymore though they were before. No idea whats going on.